requirements.txt
numpy
matplotlib
sklearn
python-mnist
